Extract ChordSheetEditorContainer
